refactor: structure and naming
