+++ src/routes/utils/member-type-controller.ts
@@ -1,5 +1,5 @@
 import { ChangeMemberTypeDTO } from '../../utils/DB/entities/DBMemberTypes';
-import { FastifyType } from '../graphql/types/fastify-type';
+import { FastifyType } from '../graphql/types/fastify';
 import { RoutesErrors } from './routes-errors';
 
 export const findMany = (fastify: FastifyType) => fastify.db.memberTypes.findMany();
